Add files via upload
Project: Compare using the Modified Tanimoto Coefficient (M.A. Fligner et al, 2002) to the Tanimoto Coefficient for selecting diverse compounds via the MinMax Selection Algorithm (Snarey et al, 1997). Analysis of the diversity of the subset chosen is based on M.A. Fligner's methods of examining feature number and average pairwise dissimilarity in the chosen subset. The Modified Tanimoto Coefficient can be used to select an initial compound, select other compounds for the subset based on the initial 
compound, and evaluate the compound selection in the final subset. Comparison of Modified Tanimotovs Tanimoto performance, however, should focus on the differences in the selection process.

Sources:
https://www.jstor.org/stable/pdf/1271255.pdf?refreqid=excelsior%3A636532b3d53dc7207a38ef884ed1afcb&ab_segments=&origin=&initiator=&acceptTC=1
https://www.sciencedirect.com/science/article/pii/S1093326398000084#BIB18